Pindah detail menjadi Perkomponen
